_IA/requirements.txt
#Pour le model
tensorflow>=2.6.0
numpy>=1.19.5
Pillow>=8.3.1

#Pour api
fastapi>=0.68.0
uvicorn>=0.15.0

# Pour chatbot
langchain==0.0.272
openai==0.27.9
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.25.0
pandas==2.0.3
faiss-cpu==1.7.4
tiktoken==0.4.0

altair==5.3.0
attrs==23.2.0
Bitdesc==0.1.4
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
imageio==2.34.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.23.2
scipy==1.13.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
streamlit==1.35.0
tenacity==8.3.0
tifffile==2024.5.22
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
Werkzeug==3.0.3
